requirement.txt
pip
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1609752033316/work
django-countries==6.1.3
django-filter==2.4.0
django-rest-passwordreset==1.1.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1609945773755/work
mysqlclient==2.0.1
Pillow==8.0.1
stripe==2.55.0

